# packages/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Datememe API
  version: 0.0.1
  description: >
    MVP contract for the favorites-based match-making app. See
    /docs/data-schema-proposal.md and /docs/project-requirements.md for the
    product and data-model rationale behind these routes.
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:3001
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Cookie-first (web), Bearer-header fallback (native apps via
        expo-secure-store). See packages/sdk/src/client.ts.
    adminAuth:
      type: http
      scheme: bearer
      description: >
        Same token as bearerAuth, plus a role check (User.role === ADMIN) —
        see plugins/security.ts's adminAuth. Used on operations restricted to
        the DATEMEME admin area.
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Not permitted
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }

    # ── Identity ──────────────────────────────────────────
...
    AdminQueueMetrics:
      type: object
      required: [metrics, oldestPending, latestFailed]
      properties:
        metrics:
          type: object
          description: 'Job count by JobStatus, e.g. { PENDING: 3, FAILED: 1 }. Only statuses with at least one job are present.'
          additionalProperties: { type: integer }
        oldestPending:
          type: array
          items: { $ref: '#/components/schemas/AdminJobSummary' }
        latestFailed:
          type: array
          items: { $ref: '#/components/schemas/AdminJobSummary' }
...
    AdminEntity:
      allOf:
        - $ref: '#/components/schemas/AdminEntityScalar'
        - type: object
          required: [mediaAssets, externalRefs]
          properties:
            mediaAssets:
              type: array
              items: { $ref: '#/components/schemas/AdminMediaAsset' }
            externalRefs:
              type: array
              items: { type: object, additionalProperties: true }
...
paths:
...
  /admin/queue:
    get:
      operationId: getQueueMetrics
      summary: Job queue health — counts by status, oldest pending, most recent failures
      tags: [admin]
      security:
        - adminAuth: []
      responses:
        '200':
          description: Job queue metrics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminQueueMetrics' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /admin/taxonomy/entities:
    get:
      operationId: getEntities
      summary: List entities, optionally filtered by type or parent — used for the taxonomy tree browser and the Moderation merge-target picker
      tags: [admin]
      security:
        - adminAuth: []
      parameters:
        - in: query
          name: entityTypeId
          schema: { type: string }
        - in: query
          name: parentId
          schema: { type: string }
          description: Pass the literal string "null" to fetch root-level entities of a type.
      responses:
        '200':
          description: Matching entities
          content:
            application/json:
              schema:
                type: object
                required: [entities]
                properties:
                  entities:
                    type: array
                    items: { $ref: '#/components/schemas/AdminEntity' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createEntity
      summary: Create a new entity value under a type (and optional parent entity)
      tags: [admin]
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminCreateEntityInput' }
      responses:
        '201':
          description: The created entity
          content:
            application/json:
              schema:
                type: object
                required: [entity]
                properties:
                  entity: { $ref: '#/components/schemas/AdminEntityScalar' }
        '400':
          description: Missing field, or parent entity's type does not match the expected hierarchy
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/taxonomy/entities/{id}:
    put:
      operationId: updateEntity
      summary: Update an entity's name, slug, parent, or status
      tags: [admin]
      security:
        - adminAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminUpdateEntityInput' }
      responses:
        '200':
          description: The updated entity
          content:
            application/json:
              schema:
                type: object
                required: [entity]
                properties:
                  entity: { $ref: '#/components/schemas/AdminEntityScalar' }
        '400':
          description: Missing field, reparent cycle detected, or new parent's type does not match the expected hierarchy
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/taxonomy/entities/generate:
    post:
      operationId: generateEntities
      summary: AI-generate candidate entity names for a category (review-and-save via the bulk endpoint, nothing is persisted here)
      tags: [admin]
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminGenerateEntitiesInput' }
      responses:
        '200':
          description: Candidate names
          content:
            application/json:
              schema:
                type: object
                required: [candidates]
                properties:
                  candidates:
                    type: array
                    items: { type: string }
        '400':
          description: Invalid input
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500':
          description: The AI generation call failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /admin/taxonomy/entities/bulk:
    post:
      operationId: bulkSaveEntities
      summary: Save a batch of reviewed candidate names as real entities (idempotent per name — upserts on slug)
      tags: [admin]
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminBulkSaveEntitiesInput' }
      responses:
        '200':
          description: Save result
          content:
            application/json:
              schema:
                type: object
                required: [success, count, entities]
                properties:
                  success: { type: boolean }
                  count: { type: integer }
                  entities:
                    type: array
                    items: { $ref: '#/components/schemas/AdminEntityScalar' }
        '400':
          description: Invalid input, or parent entity's type does not match the expected hierarchy
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/taxonomy/types:
    get:
      operationId: getEntityTypes
      summary: List all entity types (flat — types are hierarchical via parentId, not nested in the response)
      tags: [admin]
      security:
        - adminAuth: []
      responses:
        '200':
          description: All entity types
          content:
            application/json:
              schema:
                type: object
                required: [entityTypes]
                properties:
                  entityTypes:
                    type: array
                    items: { $ref: '#/components/schemas/AdminEntityType' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createEntityType
      summary: Create a new entity type
      tags: [admin]
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminCreateEntityTypeInput' }
      responses:
        '200':
          description: The created entity type
          content:
            application/json:
              schema:
                type: object
                required: [entityType]
                properties:
                  entityType: { $ref: '#/components/schemas/AdminEntityTypeScalar' }
        '400':
          description: Missing field, or parent type not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/taxonomy/types/{id}:
    put:
      operationId: updateEntityType
      summary: Update an entity type's label, slug, parent, icon, or active status
      tags: [admin]
      security:
        - adminAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminUpdateEntityTypeInput' }
      responses:
        '200':
          description: The updated entity type
          content:
            application/json:
              schema:
                type: object
                required: [entityType]
                properties:
                  entityType: { $ref: '#/components/schemas/AdminEntityTypeScalar' }
        '400':
          description: Missing field, or reparent cycle detected
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Admin: List Definitions (batch 6) ────────────────────────────────
...
  /admin/users:
    get:
      operationId: getUsers
      summary: List users, optionally filtered by search term or active plan
      tags: [admin]
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - in: query
          name: search
          schema: { type: string }
          description: Matches email, exact user id, display name, or username.
        - in: query
          name: planId
          schema: { type: string }
          description: Only users with an ACTIVE subscription to this plan.
      responses:
        '200':
          description: A page of users
          content:
            application/json:
              schema:
                type: object
                required: [users, hasMore, nextCursor]
                properties:
                  users:
                    type: array
                    items: { $ref: '#/components/schemas/AdminUserSummary' }
                  hasMore: { type: boolean }
                  nextCursor: { type: string, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/users/{id}:
    get:
      operationId: getUser
      summary: A single user's full account state, membership, and admin audit history
      tags: [admin]
      security:
        - adminAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                type: object
                required: [user, auditEvents]
                properties:
                  user: { $ref: '#/components/schemas/AdminUserDetail' }
                  auditEvents:
                    type: array
                    items: { $ref: '#/components/schemas/AdminAuditEvent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...